inject IAuthService and store userContext
