+++ src/ToDo.UI/App.xaml.cs
@@ -83,14 +83,16 @@ public sealed partial class App : Application
 	{
 		//TODO:There is a IAuthenticationService already to use it with injection
 		var auth = new AuthenticationService();
-		var userContext = await auth.Login();
+		//TODO:We need to store userContext in order to use it in the HomePage
+		var userContext = await auth.LoginAsync();
 		if (userContext.AccessToken is not null)
 		{
 			return userContext.AccessToken;
 		}
 		else
 		{
-			return string.Empty;
+			Console.WriteLine("Error: Access token is null");
+			throw new MsalClientException("", "Access token is null");
 		}
 	}
