+++ src/ToDo/Business/Services/IAuthenticationService.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using ToDo.Business.Entities;
 
 namespace ToDo.Business.Services
 {
@@ -6,6 +6,6 @@ namespace ToDo.Business.Services
 	{
 		//This will be called from WelcomeViewModel in order to get access token and user's data(email,name).
 		//because if the user tap on Close icon on the webview authentication screen the user should be redirected to WelcomePage
-		Task<UserContext> Login();
+		Task<UserContext> LoginAsync();
 	}
 }
